+++ src/ks37-window.h
@@ -8,7 +8,7 @@ G_BEGIN_DECLS
 
 G_DECLARE_FINAL_TYPE (Ks37Window, ks37_window, KS37, WINDOW, AdwApplicationWindow)
 
-void ks37_window_register_control(Ks37Window *self, uint8_t control_id, GtkWidget *widget);
+void ks37_window_register_control(Ks37Window *self, uint16_t control_id, GtkWidget *widget);
 
 G_END_DECLS
